+++ tinop_be/app/Http/Controllers/Api/TaskController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\Http\Controllers\Api;
+namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
@@ -9,11 +9,15 @@ use App\Models\Task;
 use App\Http\Resources\TaskResource;
 use App\Http\Resources\TaskCollection;
 use App\Filters\TaskFilter;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class TaskController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * @param Request $request
      * @return TaskCollection
@@ -21,13 +25,13 @@ class TaskController extends Controller
     public function index(Request $request): TaskCollection
     {
         $filter = new TaskFilter();
-
+        $query = Task::query()->with(['user', 'comments']);
         $queryItems =  $filter->transform($request);
 
         if (count($queryItems) == 0) {
             return new TaskCollection(Task::paginate());
         } else {
-            return new TaskCollection(Task::query()->where($queryItems)->paginate());
+            return new TaskCollection($query->where($queryItems)->paginate());
         }
     }
 
@@ -37,7 +41,7 @@ class TaskController extends Controller
      */
     public function show(Task $task): TaskResource
     {
-        return new TaskResource($task);
+        return new TaskResource($task->load(['user', 'comments']));
     }
 
     /**
@@ -46,24 +50,36 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $taskRequest): TaskResource
     {
-        return new TaskResource(Task::create($taskRequest->all()));
+        $task = Task::create(
+            $taskRequest->validated() + ['user_id' => auth()->id()]
+        );
+
+        return new TaskResource($task->load('user'));
     }
 
     /**
      * @param Task $task
      * @param UpdateTaskRequest $taskRequest
+     * @return TaskResource
+     * @throws AuthorizationException
      */
-    public function update(Task $task, UpdateTaskRequest $taskRequest): void
+    public function update(Task $task, UpdateTaskRequest $taskRequest): TaskResource
     {
-        $task->update($taskRequest->all());
+        $this->authorize('update', $task);
+        $task->update($taskRequest->validated());
+
+        return new TaskResource($task->fresh());
     }
 
     /**
      * @param Task $task
      * @return Response
+     * @throws AuthorizationException
      */
     public function destroy(Task $task): Response
     {
+        $this->authorize('delete', $task);
+
         $task->delete();
         return response()->noContent();
     }
